python/mosquito-server-client-messaging/requirements.txt
# pip        22.0.2
# setuptools 59.6.0
paho-mqtt==2.1.0
